+++ common/routes/account/account.connected.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import Helmet from 'react-helmet'
-import { Link } from 'react-router'
 import { StyleSheet, css } from 'aphrodite'
 import Account from '../../components/account'
 
@@ -10,10 +9,6 @@ const AccountPage = () => {
       <Helmet title='Account' />
       <h2 className={css(styles.title)}>Account Info</h2>
       <Account />
-      <Link to='account/dashboard'>
-        <div className={css(styles.button)}>View Event Dashboard</div>
-      </Link>
-      <p>show this link only if they are venues</p>
     </div>
   )
 }
@@ -26,14 +21,6 @@ const styles = StyleSheet.create({
     fontSize: 28,
     margin: '0 auto 1.5rem',
     color: '#b7b7b7'
-  },
-  button: {
-    width: '50%',
-    padding: 10,
-    borderRadius: 5,
-    textAlign: 'center',
-    backgroundColor: '#242424',
-    color: '#fff'
   }
 })
